requirements.txt
# Database-Connect MCP Server
# Core Dependencies

# FastMCP - Framework para servidores MCP
fastmcp>=0.2.0

# Conectores de Base de Datos
pymysql>=1.1.0
mysql-connector-python>=8.2.0
psycopg2-binary>=2.9.9

# Gestión de Configuración
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilidades
typing-extensions>=4.9.0
click>=8.1.7

# Seguridad
cryptography>=41.0.0

# Logging y Monitoreo
colorlog>=6.8.0

# Desarrollo y Testing (opcional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.1

# Build
build>=1.0.3
twine>=4.0.2
